Add navigation pages and dark theme toggle
